Move terminalProcessExtHostProxy, node->common
