+++ src/tsconfig.strictNullChecks.json
@@ -293,10 +293,10 @@
 		"./vs/workbench/contrib/terminal/common/terminalCommands.ts",
 		"./vs/workbench/contrib/terminal/common/terminalEnvironment.ts",
 		"./vs/workbench/contrib/terminal/common/terminalMenu.ts",
+		"./vs/workbench/contrib/terminal/common/terminalProcessExtHostProxy.ts",
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
-		"./vs/workbench/contrib/terminal/node/terminalProcessExtHostProxy.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/windowsShellHelper.ts",
 		"./vs/workbench/contrib/terminal/test/electron-browser/terminalColorRegistry.test.ts",
